paddlemix/models/mar/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.13.4
black==22.8.0
bleach==6.1.0
cbor2==5.6.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
diffusion==6.11.5
diffusion-core==0.0.71
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.12.3
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.11.5
jedi==0.19.2
jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.7
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
markupsafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
notebook==7.3.3
notebook-shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opt-einsum==3.3.0
overrides==7.7.0
packaging==25.0
paconvert==3.0.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.35.1
reprod-log==1.0.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
safetensors==0.5.3
scipy==1.10.1
send2trash==1.8.3
setuptools==75.3.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
stringcase==1.2.0
structlog==21.5.0
sympy==1.13.3
terminado==0.18.1
timm==0.9.12
tinycss2==1.2.1
tomli==2.2.1
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
triton==2.2.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.15.2